restructuring after going back to commit before
