+++ pages/artwork-info/[id].js
@@ -1,10 +1,9 @@
 import ArtworkInfo from "@/components/ArtworkInfo/index.js";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import NotesArea from "@/components/NotesArea";
 import { useState } from "react";
 
-export default function ArtworkInfoPage({ artworks }) {
+export default function ArtworkInfoPage({ artworks, onHandleTextAreaValue }) {
   const router = useRouter();
   const { id } = router.query;
   const currentArtworkInfo = artworks.items.find((item) => item.id === id);
@@ -28,8 +27,12 @@ export default function ArtworkInfoPage({ artworks }) {
       <Link href={`/`}>Go back to Artworkslist</Link>
       <Link href={`/search`}>Go back to search</Link>
       <h1>ArtworkInfoPage</h1>
-      <ArtworkInfo artwork={currentArtworkInfo} />
-      <NotesArea note={note} onNoteChange={handleNoteChange} />
+      <ArtworkInfo
+        artwork={currentArtworkInfo}
+        note={note}
+        onNoteChange={handleNoteChange}
+        onHandleTextAreaValue={onHandleTextAreaValue}
+      />
     </>
   );
 }
